03_transcribe/requirements.txt
sounddevice
numpy
webrtcvad
openai
python-dotenv
scipy